lambda_functions/aviationedge/load_dep_daily/requirements.txt
boto3==1.35.49
pandas==2.2.3
psycopg2-binary==2.9.10
pyyaml==6.0.2
tqdm==4.67.1
python-dotenv==1.0.0
pydantic==1.10.2